chore(base): Merge `rooms::normal` into `rooms`.
This patch merges the `rooms/normal.rs` file into `rooms/mod.rs`.

The name `normal` is present for historical reasons that no longer stand
today. This is no needed anymore. Let's simplify the modules.
